test_docker/requirements.txt
elasticsearch==8.14.0
requests==2.25.1
fastapi==0.111.1
pyarrow==17.0.0
python-dotenv==1.0.1
